requirements.txt
fastapi
uvicorn
langchain
mysql-connector-python
streamlit
